+++ src/main/kotlin/io/github/siyual_park/config/PreDefinedScope.kt
@@ -1,0 +1,7 @@
+package io.github.siyual_park.config
+
+object PreDefinedScope {
+    object AccessToken {
+        const val create = "create:access-token"
+    }
+}
